pyproject.toml
[tool.poetry]
name = "intelligent-agent"
version = "1.0.0"
description = "An extensible AI agent that intelligently selects and uses tools to provide personalized weekly recommendations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
python-dotenv = "^1.0.0"
openai = "^1.12.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.29.0"
aiosqlite = "^0.19.0"
aiofiles = "^23.2.0"
httpx = "^0.27.0"
pyyaml = "^6.0.0"
structlog = "^23.2.0"
python-dateutil = "^2.8.0"

# MCP - optional due to dependency conflicts
mcp = {version = "^1.12.0", optional = true}

# Optional dependencies for full features
aiomysql = {version = "^0.2.0", optional = true}
snowflake-connector-python = {version = "^3.7.0", optional = true}
google-auth = {version = "^2.25.0", optional = true}
google-auth-oauthlib = {version = "^1.2.0", optional = true}
google-auth-httplib2 = {version = "^0.2.0", optional = true}
google-api-python-client = {version = "^2.110.0", optional = true}
PyPDF2 = {version = "^3.0.0", optional = true}
python-docx = {version = "^1.1.0", optional = true}
pandas = {version = "^2.1.0", optional = true}
numpy = {version = "^1.25.0", optional = true}
scikit-learn = {version = "^1.3.0", optional = true}
notion-client = {version = "^2.2.0", optional = true}
slack-sdk = {version = "^3.26.0", optional = true}
requests = "^2.32.4"

[tool.poetry.extras]
mcp = ["mcp"]
google-drive = ["google-auth", "google-auth-oauthlib", "google-auth-httplib2", "google-api-python-client", "PyPDF2", "python-docx"]
data-analysis = ["pandas", "numpy", "scikit-learn"]
database = ["aiomysql"]
snowflake = ["snowflake-connector-python", "pandas"]
collaboration = ["notion-client", "slack-sdk"]
full = ["mcp", "google-auth", "google-auth-oauthlib", "google-auth-httplib2", "google-api-python-client", "PyPDF2", "python-docx", "pandas", "numpy", "scikit-learn", "aiomysql", "snowflake-connector-python", "notion-client", "slack-sdk"]

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-asyncio = "^0.21.0"
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
flake8 = "^6.1.0"
pre-commit = "^3.5.0"

[tool.poetry.scripts]
intelligent-agent = "src.api.main:app"
agent-cli = "cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 100
known_first_party = ["src"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "mcp.*",
    "notion_client.*",
    "slack_sdk.*",
    "google.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --strict-markers"
testpaths = [
    "tests",
]
python_files = [
    "test_*.py",
    "*_test.py",
]
python_classes = [
    "Test*",
]
python_functions = [
    "test_*",
]
asyncio_mode = "auto"
